solana_loader_v3_interface
Module instruction

Instructions for the upgradable BPF loader.

Enums

UpgradeableLoaderInstruction

Functions

close

Returns the instructions required to close a buffer account

close_any

Returns the instructions required to close program, buffer, or uninitialized account

create_buffer

Returns the instructions required to initialize a Buffer account.

deploy_with_max_program_lenDeprecated

Returns the instructions required to deploy a program with a specified maximum program length. The maximum length must be large enough to accommodate any future upgrades.

extend_program

Returns the instruction required to extend the size of a program’s executable data account

is_close_instruction

is_migrate_instruction

is_set_authority_checked_instruction

is_set_authority_instruction

is_upgrade_instruction

migrate_program

Returns the instructions required to migrate a program to loader-v4.

set_buffer_authority

Returns the instructions required to set a buffers’s authority.

set_buffer_authority_checked

Returns the instructions required to set a buffers’s authority. If using this instruction, the new authority must sign.

set_upgrade_authority

Returns the instructions required to set a program’s authority.

set_upgrade_authority_checked

Returns the instructions required to set a program’s authority. If using this instruction, the new authority must sign.

upgrade

Returns the instructions required to upgrade a program.

write

Returns the instructions required to write a chunk of program data to a buffer account.